Tikzmarknode remembers typeset picture
When a tikz picture is used in certain circumstances, such as
mathchoice, it might be processed multiple times but only typeset once.
In the node information the last processed version is remembered, but it
might have no location if it wasn't the version actually typeset.  This
commit adds a way for tikzmarknode to figure out which picture was
actually typeset and only remember the nodes associated with that
picture.  It may be of use independently on other tikzpictures.

It also makes it so that tikzmarknode respects the prefix and suffix on
node names.