Fixes for saving nodes
Primarily fixing stuff that ought to have worked but didn't, such as
handling empty parameters correctly when passing from TikZ to L3.

Cleaned up a bit of code: used the `output` technique from `spath3` for
one internal function, and clarified that a `clist` that was treated
globally actually was global.