Highlighting names frozen
In the highlighting code, the `tikzmark` names are frozen when the code
is added to the shipout hooks, so it shouldn't be passed through
`tikzmark@pp@name` again.

It's entirely possible that this breaks something else. It might be that
we actually need to set the prefix and suffix to empty token lists
instead just in case the highlighting code uses a bit more sophisticated
code that I think it does.