Rounding tikzmarks
Because tikzmark uses information from the previous run, it is possible
to end up in a loop where each the document never stabilises. It is
further possible that the actual difference is so small that the human
eye would never perceive it. To surmount that situation, this introduces
a key `tikzmark round` which gives the option of rounding the
coordinates defined by a tikzmark to a given number of decimal places
(when used via the `pic cs` coordinate system).